# server/dsl/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "trpc-agent DSL Platform API",
    "description": "REST API for the trpc-agent DSL platform. This API allows frontend applications to:\n • List available components, models, tools, and reducers\n • Validate, compile, and execute DSL workflows\n • Manage workflow lifecycle (CRUD operations)\n • Stream execution events via SSE\n\nAll endpoints use JSON for request/response unless otherwise specified.",
    "version": "1.0.0",
    "contact": {
      "name": "trpc-agent Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server"
    },
    {
      "url": "http://localhost:8090",
      "description": "API v1 endpoint (prefix /api/v1)"
    }
  ],
  "tags": [
    {
      "name": "components",
      "description": "Component registry operations"
    },
    {
      "name": "models",
      "description": "Model registry operations"
    },
    {
      "name": "tools",
      "description": "Tool registry operations"
    },
    {
      "name": "workflows",
      "description": "Workflow management (CRUD)"
    },
    {
      "name": "execution",
      "description": "Workflow execution and streaming"
    },
    {
      "name": "validation",
      "description": "DSL validation"
    }
  ],
  "paths": {
    "/api/v1/components": {
      "get": {
        "tags": ["components"],
        "summary": "List all registered components",
        "description": "Returns metadata for all components in the registry, including inputs, outputs, and configuration schema.",
        "operationId": "listComponents",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "Filter by component category",
            "schema": {
              "type": "string",
              "enum": ["llm", "tools", "retrieval", "processing", "control", "custom"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search components by name or description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of components",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComponentMetadata"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/components/{name}": {
      "get": {
        "tags": ["components"],
        "summary": "Get component details by name",
        "operationId": "getComponent",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "Component name (e.g., 'builtin.llmagent')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Component metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentMetadata"
                }
              }
            }
          },
          "404": {
            "description": "Component not found"
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "tags": ["models"],
        "summary": "List all registered models",
        "operationId": "listModels",
        "responses": {
          "200": {
            "description": "List of model names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools": {
      "get": {
        "tags": ["tools"],
        "summary": "List all registered tools",
        "operationId": "listTools",
        "responses": {
          "200": {
            "description": "List of tool metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolMetadata"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows": {
      "get": {
        "tags": ["workflows"],
        "summary": "List all workflows",
        "operationId": "listWorkflows",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workflows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowList"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["workflows"],
        "summary": "Create a new workflow",
        "operationId": "createWorkflow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workflow"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid workflow definition"
          }
        }
      }
    },
    "/api/v1/workflows/{id}": {
      "get": {
        "tags": ["workflows"],
        "summary": "Get workflow by ID",
        "operationId": "getWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "responses": {
          "200": {
            "description": "Workflow details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workflow not found"
          }
        }
      },
      "put": {
        "tags": ["workflows"],
        "summary": "Update workflow",
        "operationId": "updateWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workflow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workflow updated"
          }
        }
      },
      "delete": {
        "tags": ["workflows"],
        "summary": "Delete workflow",
        "operationId": "deleteWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "responses": {
          "204": {
            "description": "Workflow deleted"
          }
        }
      }
    },
    "/api/v1/workflows/validate": {
      "post": {
        "tags": ["validation"],
        "summary": "Validate a workflow DSL",
        "description": "Validates workflow structure, component references, and configuration without executing it.",
        "operationId": "validateWorkflow",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workflow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/schema": {
      "post": {
        "tags": ["validation"],
        "summary": "Infer workflow state schema",
        "description": "Infers state fields and usage (readers/writers) for a given workflow (view DSL). Intended for editor variable suggestions.",
        "operationId": "getWorkflowSchema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workflow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inferred state fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSchemaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/vars": {
      "post": {
        "tags": ["validation"],
        "summary": "List per-node variables for a workflow",
        "description": "Returns variables produced by each node (name/type/kind/json_schema) for editor variable pickers.",
        "operationId": "getWorkflowVars",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Workflow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Per-node variable view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVarsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/vars/node": {
      "post": {
        "tags": ["validation"],
        "summary": "List variables for a single node",
        "description": "Returns variables available when editing a single node. This is a convenience wrapper over /workflows/vars for editors that want node-scoped variable pickers.",
        "operationId": "getWorkflowNodeVars",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowVarsNodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variables for the requested node",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowVarsNode"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (e.g., malformed JSON or missing node_id)"
          },
          "404": {
            "description": "Node not found in workflow"
          }
        }
      }
    },
    "/api/v1/workflows/{id}/compile": {
      "post": {
        "tags": ["workflows"],
        "summary": "Compile a workflow to executable graph",
        "description": "Compiles the DSL workflow into an executable StateGraph. Returns compilation status and any errors.",
        "operationId": "compileWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "responses": {
          "200": {
            "description": "Compilation successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompilationResult"
                }
              }
            }
          },
          "400": {
            "description": "Compilation failed"
          }
        }
      }
    },
    "/api/v1/workflows/{id}/execute": {
      "post": {
        "tags": ["execution"],
        "summary": "Execute a workflow (non-streaming)",
        "description": "Executes the workflow and returns all events after completion.",
        "operationId": "executeWorkflow",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/{id}/execute/stream": {
      "post": {
        "tags": ["execution"],
        "summary": "Execute a workflow with SSE streaming",
        "description": "Executes the workflow and streams events via Server-Sent Events.",
        "operationId": "executeWorkflowStream",
        "parameters": [
          {
            "$ref": "#/components/parameters/workflowId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "workflowId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Workflow ID",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "ComponentMetadata": {
        "type": "object",
        "description": "Component metadata including config schema and state IO.",
        "properties": {
          "name": {
            "type": "string",
            "example": "builtin.llmagent"
          },
          "display_name": {
            "type": "string",
            "description": "Human-readable component name"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier or emoji"
          },
          "color": {
            "type": "string",
            "description": "Hex color used for UI representation"
          },
          "version": {
            "type": "string"
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterMetadata"
            }
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterMetadata"
            }
          },
          "config_schema": {
            "type": "array",
            "description": "Schema for node.config parameters",
            "items": {
              "$ref": "#/components/schemas/ParameterMetadata"
            }
          }
        }
      },
      "ParameterMetadata": {
        "type": "object",
        "description": "Parameter schema used for inputs, outputs, and config.",
        "properties": {
          "name": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Go type or logical type string"
          },
          "type_id": {
            "type": "string",
            "description": "Logical DSL type identifier (e.g., string, number, graph.messages)"
          },
          "kind": {
            "type": "string",
            "description": "Coarse-grained kind for editors: string | number | boolean | object | array | opaque"
          },
          "required": {
            "type": "boolean"
          },
          "default": {
            "description": "Default value if not provided"
          },
          "enum": {
            "type": "array",
            "items": {}
          },
          "placeholder": {
            "type": "string"
          },
          "json_schema": {
            "type": "object",
            "description": "Optional JSON Schema when this parameter is a structured object",
            "additionalProperties": true
          }
        }
      },
      "ToolMetadata": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Workflow": {
        "$ref": "../../dsl/schema/engine_dsl.schema.json#/$defs/Workflow"
      },
      "FieldUsage": {
        "type": "object",
        "description": "Describes how a particular state field is used in the workflow.",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "schema_ref": {
            "type": "string"
          },
          "json_schema": {
            "type": "object",
            "additionalProperties": true
          },
          "writers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "readers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WorkflowSchemaResponse": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldUsage"
            }
          }
        }
      },
      "WorkflowVar": {
        "type": "object",
        "properties": {
          "variable": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "json_schema": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "WorkflowVarsNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "vars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowVar"
            }
          }
        }
      },
      "WorkflowVarsNodeRequest": {
        "type": "object",
        "description": "Request body for /workflows/vars/node. Carries the full workflow draft plus the node_id currently being edited.",
        "required": ["workflow", "node_id"],
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/Workflow"
          },
          "node_id": {
            "type": "string",
            "description": "ID of the node whose available variables should be returned"
          }
        }
      },
      "WorkflowVarsResponse": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowVarsNode"
            }
          }
        }
      },
      "WorkflowResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Workflow"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "WorkflowList": {
        "type": "object",
        "properties": {
          "workflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkflowResponse"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
      "ValidationResult": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "CompilationResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "graph_id": {
            "type": "string",
            "description": "Compiled graph identifier"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExecutionRequest": {
        "type": "object",
        "properties": {
          "input": {
            "type": "object",
            "description": "Initial state/input for the workflow",
            "additionalProperties": true
          },
          "config": {
            "type": "object",
            "description": "Execution configuration",
            "properties": {
              "max_iterations": {
                "type": "integer",
                "default": 100
              },
              "timeout_seconds": {
                "type": "integer",
                "default": 300
              }
            }
          }
        }
      },
      "ExecutionResult": {
        "type": "object",
        "properties": {
          "execution_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["success", "error", "timeout"]
          },
          "final_state": {
            "type": "object",
            "description": "Final workflow state",
            "additionalProperties": true
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["node_start", "node_end", "edge", "error", "stream_chunk"]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "node_id": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        }
      }
    }
  }
}